requirements.txt
flask
psycopg
psycopg[c]
psycopg[binary]
numpy
transformers
torch
huggingface_hub
Pillow
mlx